requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
supabase==2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP Clients
httpx==0.27.0
requests==2.31.0

# AI/ML
openai==1.10.0

# Data Processing
pydantic==2.5.3

# OCR (Optional - for receipt parsing)
easyocr==1.7.0
pdf2image==1.17.0
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
